More updates

+++ tests/e2e/pkg/hello_helidon.go
@@ -4,6 +4,7 @@
 package pkg
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/onsi/gomega"
@@ -20,7 +21,11 @@ const (
 	helidonAppYaml       = "examples/hello-helidon/hello-helidon-app.yaml"
 )
 
+// DeployHelloHelidonApplication deploys the Hello Helidon example application. It accepts an optional
+// OCI Log ID that is added as an annotation on the namespace to test the OCI Logging service integration.
 func DeployHelloHelidonApplication(ociLogId string) {
+	Log(Info, "Deploy Hello Helidon Application")
+	Log(Info, fmt.Sprintf("Create namespace %s", HelloHelidonNamespace))
 	gomega.Eventually(func() (*v1.Namespace, error) {
 		nsLabels := map[string]string{
 			"verrazzano-managed": "true",
@@ -35,25 +40,34 @@ func DeployHelloHelidonApplication(ociLogId string) {
 		return CreateNamespaceWithAnnotations(HelloHelidonNamespace, nsLabels, annotations)
 	}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.BeNil())
 
+	Log(Info, "Create Hello Helidon component resource")
 	gomega.Eventually(func() error {
 		return CreateOrUpdateResourceFromFile(helidonComponentYaml)
 	}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred())
 
+	Log(Info, "Create Hello Helidon application resource")
 	gomega.Eventually(func() error {
 		return CreateOrUpdateResourceFromFile(helidonAppYaml)
 	}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred(), "Failed to create hello-helidon application resource")
 }
 
+// UndeployHelloHelidonApplication undeploys the Hello Helidon example application.
 func UndeployHelloHelidonApplication() {
-	gomega.Eventually(func() error {
-		return DeleteResourceFromFile(helidonAppYaml)
-	}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred())
+	Log(Info, "Undeploy Hello Helidon Application")
+	if exists, _ := DoesNamespaceExist(HelloHelidonNamespace); exists {
+		Log(Info, "Delete Hello Helidon application")
+		gomega.Eventually(func() error {
+			return DeleteResourceFromFile(helidonAppYaml)
+		}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred())
 
-	gomega.Eventually(func() error {
-		return DeleteResourceFromFile(helidonComponentYaml)
-	}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred())
+		Log(Info, "Delete Hello Helidon components")
+		gomega.Eventually(func() error {
+			return DeleteResourceFromFile(helidonComponentYaml)
+		}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred())
 
-	gomega.Eventually(func() error {
-		return DeleteNamespace(HelloHelidonNamespace)
-	}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred())
+		Log(Info, fmt.Sprintf("Delete namespace %s", HelloHelidonNamespace))
+		gomega.Eventually(func() error {
+			return DeleteNamespace(HelloHelidonNamespace)
+		}, helidonWaitTimeout, helidonPollingInterval).ShouldNot(gomega.HaveOccurred())
+	}
 }
